scripts/telegram-sync-python/requirements.txt
telethon==1.36.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
